# internal/keyspan: don't buffer empty span in DefragmentingIter seeks